Swapped the frame ID to an int and made it consistent everywhere. It works out at 60fps, that it will 63 years of continuous runtime...so we shouldn't run out of frame id's :p

+++ physics_debugger/Physics_Debugger/Controls/PropertyGridDisplayHelpers/ShapeListWrapper.cs
@@ -99,7 +99,7 @@ namespace physics_debugger.Controls.PropertyGridDisplayHelpers
             // Iterate the list of employees
             for (int i = 0; i < WrappedShapeIds.Count; i++)
             {
-                ShapeFrameIdPair pair = ShapeDataManager.RetrieveShapeForFrame(WrappedShapeIds[i], (uint)FrameId);
+                ShapeFrameIdPair pair = ShapeDataManager.RetrieveShapeForFrame(WrappedShapeIds[i], FrameId);
 
                 ShapePropertyDescriptor pd = new ShapePropertyDescriptor(pair.Shape, $"[{i}]");
                 pds.Add(pd);
